【更新】修复了 深度传感器 漂移问题

+++ software/rt-thread-master/bsp/stm32f40x/User/DataType.h
@@ -66,9 +66,9 @@ typedef struct
 typedef struct 
 {
 		uint8 Type;
-		int32 Depth;  	   //深度
+		float Depth;  	   //深度
 		float Temperature; //水温
-		uint32 PessureValue; 		 //压力值
+		float PessureValue; 		 //压力值
 		uint32 Init_PessureValue; //初始化采集到得压力值
 }Depth_Sensor_Type;
